Move testdata to test/
